=== polymake 4.7 ===
-- distribution --
  * support for llvm+clang 14 and gcc 12
  * support for perl 5.36

-- interfaces --
  * cdd: updated bundled version to 0.94m
  * nauty: updated bundled version to 2.7r3
  * normaliz: updated bundled version to 3.9.3
  * scip: support version 8.0.0
  * soplex: support version 6.0

-- common --
  * add disable_rules command
  * threejs:
    - update version
    - depthWrite improvements

-- matroid--
  * ehrhart polynomial for matroid polytopes
  * positroid: new rule
  * split compatibility: bug fix
  * ternary: bug fix in representation check

-- polytope --
  * convex_hull:
    - adjust labels to improve prefer behavior
    - increase weights for count rules
  * conway: add CG construction
  * johnson solids: fix J34 construction
  * SlackIdeal: type migrated to application ideal

-- topaz --
  * minimal_non_faces:
    - bug fix for column number
    - adjust preconditions for better algorithm choice
  * NON_REALIZABLE and GRASS_PLUCKER_CERTIFICATE for spheres (with holes)
  * new client poincare_sphere

+++ bundled/nauty/external/nauty/dreadnaut.c
@@ -1,7 +1,7 @@
 /*****************************************************************************
 *                                                                            *
-* This is the main file for dreadnaut() version 2.6, which is a test-bed     *
-*   for nauty() version 2.6.                                                 *
+* This is the main file for dreadnaut() version 2.7, which is a test-bed     *
+*   for nauty() version 2.7 and traces version 2.2.                          *
 *                                                                            *
 *   Subject to the copyright notice in the file COPYRIGHT.                   *
 *                                                                            *
@@ -235,7 +235,7 @@ extern void REFPROC(graph*,int*,int*,int,int*,int*,set*,int*,int,int);
 #ifndef  CANONPROC
 #define CANONPROC usercanon
 #else
-extern int CANONPROC(graph*,int*,graph*,int,int,int,int);
+extern int CANONPROC(graph*,int*,graph*,unsigned long,int,int,int);
 #endif
 
 #ifndef  INVARPROC
@@ -286,7 +286,7 @@ static void help(FILE*, int);
 static void userautom(int,int*,int*,int,int,int);
 static void usernode(graph*,int*,int*,int,int,int,int,int,int);
 static void userlevel(int*,int*,int,int*,statsblk*,int,int,int,int,int,int);
-static int usercanon(graph*,int*,graph*,int,int,int,int);
+static int usercanon(graph*,int*,graph*,unsigned long,int,int,int);
 
 static boolean options_writeautoms,options_writemarkers,
             options_digraph,options_getcanon,options_linelength;
@@ -2535,10 +2535,10 @@ userlevel(int *lab, int *ptn, int level, int *orbits, statsblk *stats,
 *****************************************************************************/
 
 static int
-usercanon(graph *g, int *lab, graph *canong, int count, int code,
+usercanon(graph *g, int *lab, graph *canong, unsigned long count, int code,
             int m, int n)
 {
     fprintf(outfile,
-	"**usercanonproc: count=%d code=%d\n",count,code);
+	"**usercanonproc: count=%lu code=%d\n",count,code);
     return 0;
 }
